refactor(chat): keep action and attachment modules bounded
